safari-534.53-branch/LayoutTests/fast/workers/storage/open-database-creation-callback-sync.html

<html>
<head>
<script src="resources/test-inputs-common.js"></script>
</head>

<body onload="runTest('resources/open-database-creation-callback-sync.js')">
This test tests openDatabaseSync()'s creation callback.
<pre id="console">
</pre>
</body>
</html>
