safari-534.53-branch/LayoutTests/fast/workers/storage/open-database-creation-callback-sync-expected.txt

This test tests openDatabaseSync()'s creation callback.
PASS: Creation callback was called.
PASS: Version set to empty string as expected.
PASS: An exception was thrown and db1Fail is null as expected.
